+++ randovania/game_description/item/ammo.py
@@ -4,6 +4,7 @@ from dataclasses import dataclass
 from frozendict import frozendict
 
 from randovania.game_description.item.item_category import ItemCategory
+from randovania.game_description.resources.location_category import LocationCategory
 from randovania.game_description.resources.pickup_entry import ResourceLock
 from randovania.game_description.resources.resource_database import ResourceDatabase
 from randovania.games.game import RandovaniaGame
@@ -16,6 +17,7 @@ class Ammo:
     name: str
     model_name: str
     items: tuple[str, ...]
+    preferred_location_category: LocationCategory
     broad_category: ItemCategory
     unlocked_by: str | None = None
     temporary: str | None = None
@@ -42,6 +44,7 @@ class Ammo:
             name=name,
             model_name=value["model_name"],
             items=frozen_lib.wrap(value["items"]),
+            preferred_location_category=LocationCategory(value["preferred_location_category"]),
             broad_category=item_categories[value["broad_category"]],
             unlocked_by=value.get("unlocked_by"),
             temporary=value.get("temporary"),
@@ -55,6 +58,7 @@ class Ammo:
         result = {
             "model_name": self.model_name,
             "items": frozen_lib.unwrap(self.items),
+            "preferred_location_category": self.preferred_location_category.value,
             "broad_category": self.broad_category.name,
             "extra": frozen_lib.unwrap(self.extra),
         }
@@ -85,5 +89,5 @@ AMMO_ITEM_CATEGORY = ItemCategory(
     name="expansion",
     long_name="Expansion",
     hint_details=("an ", "expansion"),
-    is_major=False
+    hinted_as_major=False
 )