+++ randovania/games/blank/generator/pool_creator.py
@@ -2,7 +2,8 @@ from random import Random
 
 from randovania.game_description.game_description import GameDescription
 from randovania.game_description.game_patches import GamePatches
-from randovania.game_description.resources.pickup_entry import PickupEntry, PickupModel
+from randovania.game_description.resources.location_category import LocationCategory
+from randovania.game_description.resources.pickup_entry import PickupEntry, PickupModel, PickupGeneratorParams
 from randovania.game_description.resources.resource_database import ResourceDatabase
 from randovania.games.blank.layout.blank_configuration import BlankConfiguration
 from randovania.generator.item_pool import PoolResults, pickup_creator
@@ -19,7 +20,10 @@ def create_victory_key(resource_database: ResourceDatabase):
         ),
         item_category=pickup_creator.GENERIC_KEY_CATEGORY,
         broad_category=pickup_creator.GENERIC_KEY_CATEGORY,
-        probability_offset=0.25,
+        generator_params=PickupGeneratorParams(
+            preferred_location_category=LocationCategory.MAJOR,
+            probability_offset=0.25,
+        ),
     )